Cleaned up and reorganized components js files, didn't seem to break anything.
