+++ client/routes/Home-Bestbuy-Components.js
@@ -69,61 +69,5 @@ var BestbuyIndividualResultDisplay = React.createClass({
   }
 });
 
-var BestbuyReviewsDisplay = React.createClass ({
-  render: function() {
-    var resultNodes = this.props.data.Reviews.map(function(result, index) {
-      return (
-        <BestbuyIndividualReviewDisplay
-          key={'bestbuyResult' + index}
-          title={result.title}
-          reviewer={result.reviewer}
-          comment={result.comment}
-          rating={result.rating} 
-          sku={result.sku} />
-      );
-    });
-
-    return (
-      <div className="bestbuy-reviews-display hidden">
-        <h4>Best Buy Reviews</h4>
-        <div className="row">
-          <div className="product-image-review"><img src={this.props.image} /></div>
-          <div className="product-name-review">
-            <div><strong>Product: </strong>{this.props.name}</div>
-            <div><strong>Average Rating: </strong>{this.props.data.AverageRating}</div>
-            <div><strong>Total Reviews: </strong>{this.props.data.ReviewCount}</div>
-          </div>
-        </div>
-        <hr />
-        {resultNodes}
-      </div>
-    );
-  }
-});
-
-var BestbuyIndividualReviewDisplay = React.createClass({
-  render: function() {
-    return (
-      <div className="individual-review-display">
-        <h5>
-          {this.props.title}
-        </h5>
-        <div>
-          <strong>Reviewer:</strong> {this.props.reviewer}
-        </div>
-        <div>
-          <strong>Review:</strong> {this.props.comment}
-        </div>
-        <div>
-          Rating: {this.props.rating}
-        </div>
-      </div>
-    );
-  }
-});
 
 module.exports.BestbuyRelatedResultsDisplay = BestbuyRelatedResultsDisplay;
-
-module.exports.BestbuyIndividualResultDisplay = BestbuyIndividualResultDisplay;
-
-module.exports.BestbuyReviewsDisplay = BestbuyReviewsDisplay;
